Jaswanth, Kaustav | #2104 | Fix rpm scripts to work properly during update. Remove duplicate webapp jar in etc.

+++ package/scripts/preuninstall.sh
@@ -5,8 +5,7 @@ service openmrs stop || true
 service bahmni-two-factor-auth stop || true
 
 #Copying the original web.xml before uninstalling
-cp -f /opt/bahmni-two-factor-auth/etc/openmrs/web.xml /opt/openmrs/openmrs/WEB-INF/
-rm -rf /opt/openmrs/openmrs/WEB-INF/lib/externalauth.jar
-
-
-
+if [ $1 -eq 0 ] ; then
+    cp -f /opt/bahmni-two-factor-auth/etc/openmrs/web.xml /opt/openmrs/openmrs/WEB-INF/web.xml
+    rm -rf /opt/openmrs/openmrs/WEB-INF/lib/externalauth.jar
+fi
